refactor: Route explore page via team slug (#7073)
